Rename ac_aqd to ac_calc
