+++ Dockerfile
@@ -7,7 +7,7 @@ COPY requirements.txt requirements.txt
 RUN pip install --requirement requirements.txt
 
 COPY setup.py setup.py
-COPY ac_aqd ac_aqd
+COPY ac_calc ac_calc
 
 ARG SETUP_COMMAND=develop
 RUN python setup.py $SETUP_COMMAND
